backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
redis[hiredis]==5.2.1
rq==2.1.0
prometheus-client==0.21.1
openai==1.58.1
google-generativeai==0.8.3
python-dotenv==1.0.1
